+++ frontend/pages/index.tsx
@@ -1,27 +1,32 @@
-import Image from "next/image";
-import escudoUdeA from "../public/escudoUdeA.png";
-import React from "react";
 import Link from "next/link";
-import Logo from "@/components/Logo";
+import style from "@/styles/index.module.css";
 
-const index = () => {
-  console.log(escudoUdeA);
+const Index = () => {
   return (
-    <div className="welcome flex justify-center gap-3 m-4">
-      <Logo />
-      <h1>SignAI UdeA</h1>
-      <Link href="/newlogin">
-        <button type="submit" className="primary-button">
+    <main className={style.Main}>
+      <section className={style.SectionCentral}>
+        <div>
+          <h1 className={style.Logo}>SignAI</h1>
+          <h3 className={style.Slogan}>uniendo mundos seña por seña</h3>
+        </div>
+        <Link className={style.Btn} href={"/newlogin"}>
           Ingresar
-        </button>
-      </Link>
-      <Link href="/signup">
-        <button type="reset" className="secondary-button">
-          Registrarse
-        </button>
-      </Link>
-    </div>
+        </Link>
+      </section>
+      <footer className={style.Footer}>
+        <figure className={style.WrapperImage}>
+          <img
+            className={style.Image}
+            src="udea-logo-blanco.svg"
+            alt="Logo de la Universidad de Antioquia"
+          />
+        </figure>
+        <Link className={style.Contact} href={"/"}>
+          Contáctanos
+        </Link>
+      </footer>
+    </main>
   );
 };
 
-export default index;
+export default Index;
